refactor(security): split gateway auth suites and share safe write path checks
(cherry picked from commit 83c8406f01af9a8d9a47fd0e51abab5485083f6c)
