+++ src/gateway/server.auth.default-token.test.ts
@@ -1,0 +1,9 @@
+import { describe } from "vitest";
+import { registerDefaultAuthTokenSuite } from "./server.auth.default-token.suite.js";
+import { installGatewayTestHooks } from "./server.auth.shared.js";
+
+installGatewayTestHooks({ scope: "suite" });
+
+describe("gateway server auth/connect", () => {
+  registerDefaultAuthTokenSuite();
+});
